+++ opentelemetry-otlp/CHANGELOG.md
@@ -13,6 +13,7 @@
 [#1585]: https://github.com/open-telemetry/opentelemetry-rust/pull/1585
 
 - Added `DeltaTemporalitySelector` ([#1568])
+- Add `webkpi-roots` features to `reqwest` and `tonic` backends
 
 [#1568]: https://github.com/open-telemetry/opentelemetry-rust/pull/1568